+++ dataforge-meta/src/commonMain/kotlin/hep/dataforge/values/exoticValues.kt
@@ -15,7 +15,7 @@ public class LazyParsedValue(override val string: String) : Value {
 
     override fun equals(other: Any?): Boolean = other is Value && this.parsedValue == other
 
-    override fun hashCode(): Int  = string.hashCode()
+    override fun hashCode(): Int = string.hashCode()
 }
 
 public fun String.lazyParseValue(): LazyParsedValue = LazyParsedValue(this)
@@ -23,7 +23,7 @@ public fun String.lazyParseValue(): LazyParsedValue = LazyParsedValue(this)
 /**
  * A performance optimized version of list value for doubles
  */
-public class DoubleArrayValue(override val value: DoubleArray) : Value {
+public class DoubleArrayValue(override val value: DoubleArray) : Value, Iterable<Double> {
     override val type: ValueType get() = ValueType.NUMBER
     override val number: Double get() = value.first()
     override val string: String get() = value.first().toString()
@@ -43,7 +43,9 @@ public class DoubleArrayValue(override val value: DoubleArray) : Value {
         return value.contentHashCode()
     }
 
-    override fun toString(): String = list.joinToString (prefix = "[", postfix = "]")
+    override fun toString(): String = list.joinToString(prefix = "[", postfix = "]")
+
+    override fun iterator(): Iterator<Double> = value.iterator()
 }
 
-public fun DoubleArray.asValue(): Value = if(isEmpty()) Null else DoubleArrayValue(this)
+public fun DoubleArray.asValue(): Value = if (isEmpty()) Null else DoubleArrayValue(this)